add level store and update height to uint64
1. add level store for chain
2. update type to uint64 for block height

+++ common/face/chain.go
@@ -11,19 +11,19 @@ type SnapshotReader interface {
 	HeadSnapshot() (*common.SnapshotBlock, error)
 	GetSnapshotByHashH(hashH common.HashHeight) *common.SnapshotBlock
 	GetSnapshotByHash(hash string) *common.SnapshotBlock
-	GetSnapshotByHeight(height int) *common.SnapshotBlock
+	GetSnapshotByHeight(height uint64) *common.SnapshotBlock
 	//ListSnapshotBlock(limit int) []*common.SnapshotBlock
 }
 type AccountReader interface {
 	HeadAccount(address string) (*common.AccountStateBlock, error)
 	GetAccountByHashH(address string, hashH common.HashHeight) *common.AccountStateBlock
 	GetAccountByHash(address string, hash string) *common.AccountStateBlock
-	GetAccountByHeight(address string, height int) *common.AccountStateBlock
+	GetAccountByHeight(address string, height uint64) *common.AccountStateBlock
 	//ListAccountBlock(address string, limit int) []*common.AccountStateBlock
 
 	GetAccountBySourceHash(address string, source string) *common.AccountStateBlock
 	NextAccountSnapshot() (common.HashHeight, []*common.AccountHashH, error)
-	FindAccountAboveSnapshotHeight(address string, snapshotHeight int) *common.AccountStateBlock
+	FindAccountAboveSnapshotHeight(address string, snapshotHeight uint64) *common.AccountStateBlock
 }
 
 type SnapshotWriter interface {
